moved authContext to proper location
