+++ my-app/package.json
@@ -57,7 +57,6 @@
     "@testing-library/react": "^16.0.1",
     "@types/axios": "^0.14.0",
     "@types/jest": "^29.5.13",
-    "@types/jsonwebtoken": "^9.0.7",
     "@types/js-cookie": "^3.0.6",
     "@types/jsonwebtoken": "^9.0.7",
     "@types/node": "20.12.7",
